Added cartesian coordinate to voxel cell functions
Functions to find the index  value of the one dimensional  storage vector  that  any 2D or 3D coordinate data for a point P in 2D, 3D space would be assigned to that corresponds to the area or volume of a voxel that P would be designated to occupy.

If P is outside the boundaries of the 2D hex grid or 3D voxel matrix, a negative or invalid index number is returned.